modified:   src/app/helper.c 	modified:   src/app/helper.h

+++ src/app/helper.c
@@ -3,7 +3,7 @@
 
 /* Usable functions */
 
-size_t		stringLengthHelper(string s){
+long int  stringLengthHelper(string s){
   string	bfs;
 
   for (bfs = s; *bfs != '\0'; ++bfs);
