+++ src/app/helper.h
@@ -1,4 +1,4 @@
-size_t  stringLengthHelper(string s);
-void	charPrintHelper(const char c);
-void    stringPrintHelper(const string s);
-void    integerPrintHelper(int n);
+long int    stringLengthHelper(string s);
+void	    charPrintHelper(const char c);
+void        stringPrintHelper(const string s);
+void        integerPrintHelper(int n);
